Happy to Help?

A systematic review and meta-analysis of the effects of performing acts of kindness on the well-being of the actor.

Usage

data(curry)

Format

A data.frame with 56 rows and 18 columns.

Source

doi:10.1016/j.jesp.2018.02.014

Details

study_id	factor	Unique identifier of the study

effect_id	integer	Unique identifier of the effect size

d	numeric	Standardized mean difference between the control group and intervention group

vi	numeric	Variance of the effect size

n1i	numeric	Number of participants in the intervention group

n1c	numeric	Number of participants in the control group

sex	numeric	Percentage of male participants

age	numeric	Mean age of participants

location	character	Geographical location of the study

donor	character	From what population did the donors (helpers) originate?

donorcode	factor	From what population did the donors (helpers) originate? Dichotomized to Anxious or Typical

interventioniv	character	Description of the intervention / independent variable

interventioncode	factor	Description of the intervention / independent variable, categorized to Acts of Kindness, Prosocial Spending, or Other

control	character	Description of the control condition

controlcode	factor	Description of the control condition, categorized to Neutral Activity, Nothing, or Self Help (performing a kind act for oneself)

recipients	character	Who were the recipients of the act of kindness?

outcomedv	character	What was the outcome, or dependent variable, of the study?

outcomecode	factor	What was the outcome, or dependent variable, of the study? Categorized into Happiness, Life Satisfaction, PN Affect (positive or negative), and Other

References

Curry, O. S., Rowland, L. A., Van Lissa, C. J., Zlotowitz, S., McAlaney, J., & Whitehouse, H. (2018). Happy to help? A systematic review and meta-analysis of the effects of performing acts of kindness on the well-being of the actor. Journal of Experimental Social Psychology, 76, 320-329. doi:10.1016/j.jesp.2018.02.014