Does training matter? A meta-analysis of caregiver training studies

A review of 17 experimental studies published between 1980 and 2005 on the effect of specialized training on the competency of caregivers in childcare.

data(fukkink_lont)

Format

A data.frame with 78 rows and 30 columns.

Source

doi:10.1016/j.ecresq.2007.04.005

Details

id_exp	integer	Unique identifier of the study

yi	numeric	Standardized mean difference between the control group and

vi	numeric	Variance of the effect size

Journal	factor	Publication type (scientific journal or other publications)

Setting	factor	Setting (center-based care or family daycare)

Integrated	factor	Whether the training was integrated into childcare practice

Supervision	factor	Whether supervision was part of the training

Scope	factor	Scope of the training (narrow or broad)

Location	factor	Location of the training (one-site or multi-site)

Curriculum	factor	Fixed curriculum

Control	factor	Alternative treatment for control group

Assignment	factor	Random assignment or matching (at the level of the individual caregiver or childcare center)

Train_Knowledge	factor	Explicit focus on knowledge

Train_Skills	factor	Explicit focus on skills

Train_Attitude	factor	Explicit focus on attitude

Video	factor	Use of video feedback

Design	factor	Single group, or two-group experimental design

Pre_Post	factor	Pretest/posttest design (yes/no)

Blind	factor	Was a blinding procedure used?

Attrition	numeric	Attrition from the experimental condition (percentage)

Pretest_es	numeric	Pre-test effect size

Self_report	factor	Self-report measures of caregiver competencies versus ‘objective’ test or observation by independent observer

DV_Knowledge	factor	Test focused on knowledge

DV_Skills	factor	Test focused skills

DV_Attitude	factor	Test focused on attitudes

DV_Aligned	factor	Test aligned with the content of the training (yes/no)

Two_group_design	factor	Single group, or two-group experimental design

Trainee_Age	numeric	Trainees’ age

Trainee_Experience	numeric	Trainees’ working experience

n_total	integer	Total n at post-test

References

Fukkink, R. G., & Lont, A. (2007). Does training matter? A meta-analysis and review of caregiver training studies. Early childhood research quarterly, 22(3), 294-311. doi:10.1016/j.ecresq.2007.04.005